fix: rearrange tests
